requirements.txt
ultralytics==8.0.230
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
numpy==1.26.4
albumentations>=1.3.0
opencv-python>=4.7.0
tqdm
pyyaml
